KOGITO-3659 - Remove duplicate services for SW GH example
