--- serverless-workflow-github-showcase/pr-checker-workflow/src/main/java/org/kogito/examples/sw/github/workflow/GitHubServiceFrontend.java
@@ -1,29 +1,0 @@
-/**
- *  Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.kogito.examples.sw.github.workflow;
-
-import javax.enterprise.context.ApplicationScoped;
-
-/**
- * Placeholder for the SW file duo to codegen bug.
- *
- * JIRA: https://issues.redhat.com/browse/KOGITO-3659
- */
-@ApplicationScoped
-public class GitHubServiceFrontend extends AbstractGitHubService {
-
-}
