+++ serverless-workflow-github-showcase/pr-checker-workflow/src/main/resources/handle-frontend.sw.json
@@ -7,7 +7,7 @@
       "name": "AddLabels",
       "type": "service",
       "metadata": {
-        "interface": "org.kogito.examples.sw.github.workflow.GitHubServiceFrontend",
+        "interface": "org.kogito.examples.sw.github.workflow.GitHubService",
         "operation": "addLabels"
       }
     },
@@ -15,7 +15,7 @@
       "name": "AddReviewers",
       "type": "service",
       "metadata": {
-        "interface": "org.kogito.examples.sw.github.workflow.GitHubServiceFrontend",
+        "interface": "org.kogito.examples.sw.github.workflow.GitHubService",
         "operation": "addReviewers"
       }
     }
